Moved distro map elements to objects
